feat(web): initial user profile page
