+++ apps/web/src/pages/settings/index.ts
@@ -1,1 +1,2 @@
 export * from './ApiKeysPage';
+export * from './user-profile-page';
