+++ apps/web/src/pages/settings/user-profile-page/UserProfilePage.tsx
@@ -1,0 +1,9 @@
+import { SettingsPageContainer } from '../SettingsPageContainer';
+
+export const UserProfilePage = () => {
+  return (
+    <SettingsPageContainer title="User profile">
+      <div>TODO: replace with page content</div>
+    </SettingsPageContainer>
+  );
+};
