+++ apps/web/src/pages/settings/user-profile-page/index.ts
@@ -1,0 +1,1 @@
+export * from './UserProfilePage';
